Dissolve -util in favor of -mode and new -exec
The lean4-exec.el feature has been written from scratch.  Its logic is
simple: For Elan, Lake and Lean respectively, there is a hook of
functions which may return a path to the respective program.  Well,
there are some more details to it but its great at being minimalist
but scalable.